#26 refactor preferences and use __package__
