+++ save_load/__init__.py
@@ -1,2 +1,3 @@
+from .pref_access import get_user_preferences
 from .reload import reload_issues, reload_rules
 from .save_load_util import does_config_exist, get_config_filepath, import_lint_rules, save_external_rules
